fix:adapt to new ROM scroll data
<!-- PLEASE REMOVE COMMENTS BEFORE MERGING TO AVOID INCLUDING THEM IN GIT HISTORY -->
## Summary
<!--
Briefly describe the purpose of the MR, for example:

This PR adds support for rounded corners for components in C-API architecture. 
-->
New ROM scroll offset data change from `i32` to `f32`, also change the `onScroll` event interval threshold accordingly.
## Changes
<!-- List main changes in bullet points that can be included in the changelog. Those changes should be written for
application developers and react third-party library authors. Internal changes shouldn't be listed. Please start each bullet point with:
- added
- fixed
- deprecated
- removed

For breaking changes, use "[breaking]" prefix, for example:
- [breaking] removed something
-->

## Test Plan
<!-- Steps on how the changes were tested. E.g.,
1. Open tester app.
2. Run "ViewTest::render square with rounded corners with different radii (left/right)". -->

1. Open tester app.
2. 


## Checklist
- [ ] Project compiles and the tester application works.

## Issue Links
<!-- Link related or resolved issues. E.g.,
Related: #123 
Closes: #456
-->

See merge request rnoh/react-native-harmony!1155

+++ tester/harmony/react_native_openharmony/src/main/cpp/RNOH/arkui/ScrollNode.cpp
@@ -56,8 +56,8 @@ void ScrollNode::onNodeEvent(
 facebook::react::Point ScrollNode::getScrollOffset() const {
   auto item = NativeNodeApi::getInstance()->getAttribute(
       m_nodeHandle, NODE_SCROLL_OFFSET);
-  facebook::react::Float x = item->value[0].i32;
-  facebook::react::Float y = item->value[1].i32;
+  facebook::react::Float x = item->value[0].f32;
+  facebook::react::Float y = item->value[1].f32;
   return Point{x, y};
 }
